+++ x/staking/keeper/keeper_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/codec/address"
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/runtime"
 	"github.com/cosmos/cosmos-sdk/testutil"
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
@@ -181,6 +182,13 @@ func getUBDKey(delAddr sdk.AccAddress, valAddr sdk.ValAddress) []byte {
 	return append(append(unbondingDelegationKey, addresstypes.MustLengthPrefix(delAddr)...), addresstypes.MustLengthPrefix(valAddr)...)
 }
 
+// getValidatorKey creates the key for the validator with address
+// VALUE: staking/Validator
+func getValidatorKey(operatorAddr sdk.ValAddress) []byte {
+	validatorsKey := []byte{0x21}
+	return append(validatorsKey, addresstypes.MustLengthPrefix(operatorAddr)...)
+}
+
 func (s *KeeperTestSuite) TestSrcRedelegationsMigrationToColls() {
 	s.SetupTest()
 
@@ -296,7 +304,66 @@ func (s *KeeperTestSuite) TestUnbondingDelegationsMigrationToColls() {
 		},
 		"d03ca412f3f6849b5148a2ca49ac2555f65f90b7fab6a289575ed337f15c0f4b",
 	)
+	s.Require().NoError(err)
+}
 
+func (s *KeeperTestSuite) TestValidatorsMigrationToColls() {
+	s.SetupTest()
+	pkAny, err := codectypes.NewAnyWithValue(PKs[0])
+	s.Require().NoError(err)
+
+	_, valAddrs := createValAddrs(100)
+
+	err = testutil.DiffCollectionsMigration(
+		s.ctx,
+		s.key,
+		100,
+		func(i int64) {
+			val := stakingtypes.Validator{
+				OperatorAddress:   valAddrs[i].String(),
+				ConsensusPubkey:   pkAny,
+				Jailed:            false,
+				Status:            stakingtypes.Bonded,
+				Tokens:            sdk.DefaultPowerReduction,
+				DelegatorShares:   math.LegacyOneDec(),
+				Description:       stakingtypes.Description{},
+				UnbondingHeight:   int64(0),
+				UnbondingTime:     time.Unix(0, 0).UTC(),
+				Commission:        stakingtypes.NewCommission(math.LegacyZeroDec(), math.LegacyZeroDec(), math.LegacyZeroDec()),
+				MinSelfDelegation: math.ZeroInt(),
+			}
+			valBz := stakingtypes.MustMarshalValidator(s.cdc, &val)
+			// legacy Set method
+			s.ctx.KVStore(s.key).Set(getValidatorKey(valAddrs[i]), valBz)
+		},
+		"6a8737af6309d53494a601e900832ec27763adefd7fe8ff104477d8130d7405f",
+	)
+	s.Require().NoError(err)
+
+	err = testutil.DiffCollectionsMigration(
+		s.ctx,
+		s.key,
+		100,
+		func(i int64) {
+			val := stakingtypes.Validator{
+				OperatorAddress:   valAddrs[i].String(),
+				ConsensusPubkey:   pkAny,
+				Jailed:            false,
+				Status:            stakingtypes.Bonded,
+				Tokens:            sdk.DefaultPowerReduction,
+				DelegatorShares:   math.LegacyOneDec(),
+				Description:       stakingtypes.Description{},
+				UnbondingHeight:   int64(0),
+				UnbondingTime:     time.Unix(0, 0).UTC(),
+				Commission:        stakingtypes.NewCommission(math.LegacyZeroDec(), math.LegacyZeroDec(), math.LegacyZeroDec()),
+				MinSelfDelegation: math.ZeroInt(),
+			}
+
+			err := s.stakingKeeper.SetValidator(s.ctx, val)
+			s.Require().NoError(err)
+		},
+		"6a8737af6309d53494a601e900832ec27763adefd7fe8ff104477d8130d7405f",
+	)
 	s.Require().NoError(err)
 }